requirements.txt
requests
cryptos
tenacity
dacite
base58
ruamel.yaml